Code quality improvements and cleanup.

+++ elements/src/main/java/com/manydesigns/elements/util/MemoryUtil.java
@@ -54,9 +54,9 @@ public class MemoryUtil {
         }
         int log = (int) Math.log10(bytes);
         int pos = log / 3;
-        double scaled = bytes / Math.pow(10, pos * 3);
+        double scaled = ((double) bytes) / Math.pow(10, pos * 3);
         long rounded = Math.round(scaled);
-        double rounded2 = (double)Math.round(scaled * 10) / 10d;
+        double rounded2 = Math.round(scaled * 10) / 10d;
 
         if (pos == 0) {
             return MessageFormat.format("{0,number,0}{1}",
@@ -64,7 +64,7 @@ public class MemoryUtil {
         } else if (rounded >= 1000) {
             scaled = scaled / 1000;
             pos = pos + 1;
-            rounded2 = (double)Math.round(scaled * 10) / 10d;
+            rounded2 = Math.round(scaled * 10) / 10d;
             return MessageFormat.format("{0,number,0.0}{1}",
                     rounded2, suffixes[pos]);
         } else if (rounded2 < 10) {
